+++ src/main/java/io/apiman/cli/core/api/model/ServiceConfig.java
@@ -20,6 +20,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -49,6 +50,10 @@ public class ServiceConfig {
         this.endpoint = endpoint;
         this.endpointType = endpointType;
         this.publicService = publicService;
+        this.gateways = gateways;
+    }
+
+    public void setGateways(ArrayList<ApiGateway> gateways) {
         this.gateways = gateways;
     }
 }