+++ src/main/java/io/apiman/cli/core/common/action/ModelCreateAction.java
@@ -39,7 +39,7 @@ public abstract class ModelCreateAction<M, A> extends AbstractModelAction<M, A> 
     }
 
     @Override
-    protected void performAction(CmdLineParser parser) throws ActionException {
+    public void performAction(CmdLineParser parser) throws ActionException {
         LOGGER.debug("Creating {}", this::getModelName);
 
         ApiUtil.invokeAndCheckResponse(() -> {