+++ src/main/java/io/apiman/cli/core/declarative/model/DeclarativeApi.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,56 +14,53 @@
  * limitations under the License.
  */
 
-package io.apiman.cli.core.api.model;
+package io.apiman.cli.core.declarative.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import io.apiman.cli.core.api.model.Api;
+
+import java.util.List;
 
 /**
- * Models an API.
+ * Declarative API representation.
  *
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Api {
+public class DeclarativeApi extends Api {
     @JsonProperty
-    private String name;
+    private boolean published;
 
     @JsonProperty
-    private String description;
+    private DeclarativeApiConfig config;
 
     @JsonProperty
-    private String organizationName;
+    private List<DeclarativePolicy> policies;
 
-    @JsonProperty
-    private String initialVersion;
-
-    @JsonProperty
-    private String version;
-
-    @JsonProperty
-    private String status;
-
-    public Api() {
+    public boolean isPublished() {
+        return published;
     }
 
-    public Api(String name, String description, String initialVersion) {
-        this.name = name;
-        this.description = description;
-        this.initialVersion = initialVersion;
+    public void setPublished(boolean published) {
+        this.published = published;
     }
 
-    public String getName() {
-        return name;
+    public DeclarativeApiConfig getConfig() {
+        return config;
     }
 
-    public String getInitialVersion() {
-        return initialVersion;
+    public void setConfig(DeclarativeApiConfig config) {
+        this.config = config;
     }
 
-    public String getStatus() {
-        return status;
+    public List<DeclarativePolicy> getPolicies() {
+        return policies;
+    }
+
+    public void setPolicies(List<DeclarativePolicy> policies) {
+        this.policies = policies;
     }
 }