+++ src/main/java/io/apiman/cli/core/declarative/model/DeclarativePolicy.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,56 +14,41 @@
  * limitations under the License.
  */
 
-package io.apiman.cli.core.api.model;
+package io.apiman.cli.core.declarative.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.util.Map;
+
 /**
- * Models an API.
+ * Declarative policy representation.
  *
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Api {
+public class DeclarativePolicy {
     @JsonProperty
     private String name;
 
     @JsonProperty
-    private String description;
-
-    @JsonProperty
-    private String organizationName;
-
-    @JsonProperty
-    private String initialVersion;
-
-    @JsonProperty
-    private String version;
-
-    @JsonProperty
-    private String status;
-
-    public Api() {
-    }
-
-    public Api(String name, String description, String initialVersion) {
-        this.name = name;
-        this.description = description;
-        this.initialVersion = initialVersion;
-    }
+    private Map<String, Object> config;
 
     public String getName() {
         return name;
     }
 
-    public String getInitialVersion() {
-        return initialVersion;
+    public void setName(String name) {
+        this.name = name;
     }
 
-    public String getStatus() {
-        return status;
+    public Map<String, Object> getConfig() {
+        return config;
+    }
+
+    public void setConfig(Map<String, Object> config) {
+        this.config = config;
     }
 }