+++ src/main/java/io/apiman/cli/core/declarative/model/DeclarativeSystem.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,35 +14,42 @@
  * limitations under the License.
  */
 
-package io.apiman.cli.core.org.model;
+package io.apiman.cli.core.declarative.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import io.apiman.cli.core.plugin.model.Plugin;
+
+import java.util.List;
 
 /**
- * Models an organisation.
+ * Declarative system configuration.
  *
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
-@JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Org {
+@JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
+public class DeclarativeSystem {
     @JsonProperty
-    private String name;
+    private List<DeclarativeGateway> gateways;
 
     @JsonProperty
-    private String description;
+    private List<Plugin> plugins;
 
-    public Org() {
+    public List<DeclarativeGateway> getGateways() {
+        return gateways;
     }
 
-    public Org(String name, String description) {
-        this.name = name;
-        this.description = description;
+    public void setGateways(List<DeclarativeGateway> gateways) {
+        this.gateways = gateways;
     }
 
-    public String getName() {
-        return name;
+    public List<Plugin> getPlugins() {
+        return plugins;
+    }
+
+    public void setPlugins(List<Plugin> plugins) {
+        this.plugins = plugins;
     }
 }