+++ src/main/java/io/apiman/cli/util/ApiUtil.java
@@ -16,10 +16,15 @@
 
 package io.apiman.cli.util;
 
+import com.google.common.io.BaseEncoding;
 import com.google.common.io.CharStreams;
 import io.apiman.cli.exception.ActionException;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import retrofit.RestAdapter;
 import retrofit.RetrofitError;
 import retrofit.client.Response;
+import retrofit.converter.JacksonConverter;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -27,12 +32,16 @@ import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.util.function.Supplier;
 
+import static io.apiman.cli.util.MappingUtil.JSON_MAPPER;
+
 /**
  * Shared API utility methods.
  *
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
 public class ApiUtil {
+    private static final Logger LOGGER = LogManager.getLogger(ApiUtil.class);
+
     public static void invokeAndCheckResponse(Supplier<Response> request) throws ActionException {
         invokeAndCheckResponse(HttpURLConnection.HTTP_OK, request);
     }
@@ -70,5 +79,29 @@ public class ApiUtil {
 
         throw new ActionException("HTTP " + response.getStatus() + " "
                 + response.getReason() + " but expected " + expectedStatus + ":\n" + body);
+    }
+
+    /**
+     * @param clazz        the Class for which to build a client
+     * @param username     the management API username
+     * @param password     the management API password
+     * @param debugLogging whether debug logging should be enabled
+     * @param <T>          the API interface
+     * @return an API client for the given Class
+     */
+    public static <T> T buildApiClient(Class<T> clazz, String endpoint, String username, String password, boolean debugLogging) {
+        final RestAdapter.Builder builder = new RestAdapter.Builder() //
+                .setConverter(new JacksonConverter(JSON_MAPPER))
+                .setEndpoint(endpoint)
+                .setRequestInterceptor(request -> {
+                    final String credentials = String.format("%s:%s", username, password);
+                    request.addHeader("Authorization", "Basic " + BaseEncoding.base64().encode(credentials.getBytes()));
+                });
+
+        if (debugLogging) {
+            builder.setLogLevel(RestAdapter.LogLevel.FULL);
+        }
+
+        return builder.build().create(clazz);
     }
 }