+++ src/main/java/io/apiman/cli/util/OptionalConsumer.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2016 Pete Cornish
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.apiman.cli.util;
+
+import java.util.Optional;
+import java.util.function.Consumer;
+
+/**
+ * Functional convenience methods.
+ *
+ * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
+ */
+public class OptionalConsumer<T> {
+    private Optional<T> optional;
+
+    private OptionalConsumer(Optional<T> optional) {
+        this.optional = optional;
+    }
+
+    public static <T> OptionalConsumer<T> of(Optional<T> optional) {
+        return new OptionalConsumer<>(optional);
+    }
+
+    public OptionalConsumer<T> ifPresent(Consumer<T> c) {
+        optional.ifPresent(c);
+        return this;
+    }
+
+    public OptionalConsumer<T> ifNotPresent(Runnable r) {
+        if (!optional.isPresent()) {
+            r.run();
+        }
+        return this;
+    }
+}