+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- N/A
+
+## [0.1.5] - 2016-01-23
+### Added
+- Declarative API environment definitions (using YAML or JSON) with the new `apiman apply` command
+- Declarations can use placeholders, passed from the command line with the `-P key=value` or `-P "key=value"` syntax
 
 ## [0.1.4] - 2016-01-11
 ### Added